+++ packages/scenes/src/index.ts
@@ -64,7 +64,7 @@ export {
   type SceneAppDrilldownView,
   type SceneAppRoute,
 } from './components/SceneApp/types';
-export { SceneApp } from './components/SceneApp/SceneApp';
+export { SceneApp, useSceneApp } from './components/SceneApp/SceneApp';
 export { SceneAppPage } from './components/SceneApp/SceneAppPage';
 export { SceneReactObject } from './components/SceneReactObject';
 export { SceneObjectRef } from './core/SceneObjectRef';