+++ packages/scenes/src/querying/layers/annotations/AnnotationsDataLayer.ts
@@ -6,9 +6,10 @@ import { sceneGraph } from '../../../core/sceneGraph';
 import { SceneDataLayerProvider, SceneTimeRangeLike, SceneDataLayerProviderState } from '../../../core/types';
 import { getDataSource } from '../../../utils/getDataSource';
 import { getMessageFromError } from '../../../utils/getMessageFromError';
+import { writeSceneLog } from '../../../utils/writeSceneLog';
 import { SceneDataLayerBase } from '../SceneDataLayerBase';
 import { AnnotationQueryResults, executeAnnotationQuery } from './standardAnnotationQuery';
-import { postProcessQueryResult } from './utils';
+import { dedupAnnotations, postProcessQueryResult } from './utils';
 
 interface AnnotationsDataLayerState extends SceneDataLayerProviderState {
   query: AnnotationQuery;
@@ -22,10 +23,13 @@ export class AnnotationsDataLayer
   public topic = DataTopic.Annotations;
 
   public constructor(initialState: AnnotationsDataLayerState) {
-    super({
-      isEnabled: true,
-      ...initialState,
-    });
+    super(
+      {
+        isEnabled: true,
+        ...initialState,
+      },
+      ['query']
+    );
   }
 
   public onEnable(): void {
@@ -41,6 +45,7 @@ export class AnnotationsDataLayer
   }
 
   public runLayer() {
+    writeSceneLog('AnnotationsDataLayer', 'run layer');
     const timeRange = sceneGraph.getTimeRange(this);
     this.runWithTimeRange(timeRange);
   }
@@ -52,10 +57,16 @@ export class AnnotationsDataLayer
       this.querySub.unsubscribe();
     }
 
+    if (sceneGraph.hasVariableDependencyInLoadingState(this)) {
+      writeSceneLog('AnnotationsDataLayer', 'Variable dependency is in loading state, skipping query execution');
+      this.setState({ _isWaitingForVariables: true });
+      return;
+    }
+
     try {
       const ds = await this.resolveDataSource(query);
 
-      const queryExecution = executeAnnotationQuery(ds, timeRange, query).pipe(
+      const queryExecution = executeAnnotationQuery(ds, timeRange, query, this).pipe(
         map((events) => {
           const stateUpdate = this.processEvents(query, events);
           return stateUpdate;
@@ -87,7 +98,9 @@ export class AnnotationsDataLayer
   }
 
   protected processEvents(query: AnnotationQuery, events: AnnotationQueryResults) {
-    const processedEvents = postProcessQueryResult(query, events.events || []);
+    let processedEvents = postProcessQueryResult(query, events.events || []);
+    processedEvents = dedupAnnotations(processedEvents);
+
     const stateUpdate = { ...emptyPanelData, state: events.state };
     const df = arrayToDataFrame(processedEvents);
     df.meta = {