+++ packages/scenes/src/querying/layers/annotations/utils.ts
@@ -4,7 +4,7 @@
 
 import { AnnotationEvent, AnnotationQuery, DataSourceApi } from '@grafana/data';
 import { config } from '@grafana/runtime';
-import { cloneDeep } from 'lodash';
+import { cloneDeep, concat, every, find, groupBy, head, map, partition } from 'lodash';
 
 const legacyRunner = [
   'prometheus',
@@ -63,4 +63,30 @@ export function postProcessQueryResult(annotation: AnnotationQuery, results: Ann
   });
 
   return processed;
+}
+
+export function dedupAnnotations(annotations: any) {
+  let dedup = [];
+
+  // Split events by annotationId property existence
+  const events = partition(annotations, 'id');
+
+  const eventsById = groupBy(events[0], 'id');
+  dedup = map(eventsById, (eventGroup) => {
+    if (eventGroup.length > 1 && !every(eventGroup, isPanelAlert)) {
+      // Get first non-panel alert
+      return find(eventGroup, (event) => {
+        return event.eventType !== 'panel-alert';
+      });
+    } else {
+      return head(eventGroup);
+    }
+  });
+
+  dedup = concat(dedup, events[1]);
+  return dedup;
+}
+
+function isPanelAlert(event: { eventType: string }) {
+  return event.eventType === 'panel-alert';
 }