+++ packages/scenes/src/variables/adhoc/AdHocFilterSeparator.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export function AdHocFilterSeparator() {
+  return <div>{/** AND */}</div>;
+}