+++ packages/scenes/src/variables/adhoc/AdHocFilterValue.tsx
@@ -15,7 +15,7 @@ export function AdHocFilterValue({ filter, placeHolder, model }: Props) {
   const loadValues = () => model._getValuesFor(filter);
 
   return (
-    <div className="gf-form" data-testid="AdHocFilterValue-value-wrapper">
+    <div data-testid="AdHocFilterValue-value-wrapper">
       <SegmentAsync
         className="query-segment-value"
         disabled={model.state.readOnly}